Feat: Implementing the whole fetching pokemon effect
The useEffect at the start of the App runs one time. It fetches the link with a random number. Since it returns an array of all the pokemon urls, a second fetch is needed to take the front_sprite of the image, that gets pushed into the pokemon array that will be used later