Fix handleCardsClick function
The function seems pretty hard to analyze but it's easy. If the array of clicked pokemons is empty, just add one to score and randomize the order of pokemons. Then it automatically changes the score and clicked pokemons. This because the return statement of forEach does include the code after.